% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{stock_fundamentals}
\alias{stock_fundamentals}
\title{Stock fundamental data.}
\format{
A data frame with 8702 rows and 10 variables:
\describe{
\item{date_stamp}{month end dates}
\item{sector}{the industry sector to whcih the stock belongs}
\item{ticker}{the ticker symbol identifying the company}
\item{log_pb}{the log of the book to price ratio}
\item{mkt_cap}{the stocks market capitalisation}
\item{reo}{the stocks return on equity for the trailing 4 quarters}
\item{total_assets}{the stocks total assets at the most recent available reporting date}
\item{total_equity}{the stocks total equity at the most recent available reporting date}
\item{total_equity_cln}{should the stocks have negative total equity, 10\% of total assets}
\item{leverage}{total assets divided by total liabilities}
...
}
}
\source{
\url{https://www.sec.gov/edgar/searchedgar/companysearch.html}
}
\usage{
stock_fundamentals
}
\description{
A dataset containing various stock level characteristics for approximately 700 companies for each month in 2017
}
\keyword{datasets}
