% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipc_suv.R
\name{suv}
\alias{suv}
\title{Standardised unexpected volume}
\usage{
suv(df)
}
\arguments{
\item{df}{A data frame containing columns 'date_stamp', 'pos', 'neg'.}
}
\value{
A tibble containing a date stamp and standardised unexpected volume.
}
\description{
Implementation as per:
Garfinkel, Jon A. and Sokobin, Jonathan, Volume, Opinion Divergence and Returns: A Study of Post-Earnings Announcement Drift (November 2003). AFA 2004 San Diego Meetings, Available at SSRN: https://ssrn.com/abstract=280913 or http://dx.doi.org/10.2139/ssrn.280913
}
