% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_segmentation.R
\name{ts_segmentation}
\alias{ts_segmentation}
\title{Time series histogram and shading}
\usage{
ts_segmentation(df1, df2, col1, col2)
}
\arguments{
\item{df1}{A dataframe containing the following columns:
\itemize{
\item date
\item market index values (labelled "close")
\item monthly forward market returns (labelled "fwd_rtn_m")
\item an indicator time series to be plotted and categorised into bins representing specific level and change values
}}

\item{df2}{A dataframe containing the time series to shade the S&P500 plot}

\item{col1}{A column name in df1 representing time series to derive the multiple binary indicators}

\item{col2}{either, level - split time series into 3 levels, or both  - split time series into 3 levels and 6 month change}
}
\value{
A tibble containing a date stamp and standardised unexpected volume.
}
\description{
Make density plot of subsequent returns conditioned on multiple binary indicators derived from reference time series
}
