% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xsect_scale.R
\name{xsect_scale}
\alias{xsect_scale}
\title{Cross sectional scale specific columns in a data frame}
\usage{
xsect_scale(df, preds)
}
\arguments{
\item{df}{A data frame containing columns 'date_stamp', 'date_char', 'symbol' and 'fwd_rtn'.}

\item{preds}{A list or character vector of columns names included in df.}
}
\value{
A tibble containing specified input columns and the columns specified with 'preds' on a standardised basis.
}
\description{
Cross sectional scale specific columns in a data frame
}
