% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipc_suv.R
\name{suv}
\alias{suv}
\title{Standardised unexpected volume}
\usage{
suv(df)
}
\arguments{
\item{df}{A data frame containing columns 'date_stamp', 'pos', 'neg'}
}
\value{
A tibble containing a date stamp and standardised unexpected volume
}
\description{
Standardised unexpected volume
}
