% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_segmentation.R
\name{ts_segmentation}
\alias{ts_segmentation}
\title{Time series histogram and shading}
\usage{
ts_segmentation(
  df1,
  date_idx,
  invest_series,
  cndtn_series,
  bin_method,
  lb = 6,
  pc = 0.2,
  fr = -0.05
)
}
\arguments{
\item{df1}{A dataframe containing the following columns:
\itemize{
\item date
\item a times series for assessment (to be referenced by the argument "invest_series")
\item monthly forward market returns (labelled "fwd_rtn_m")
\item an indicator time series (to be referenced by the argument "cndtn_series") for plotting and categorisation into bins representing specific level and change values
}}

\item{date_idx}{The column in df1 representing the date index}

\item{invest_series}{A column in df1 representing the time series for which returns are to be assessed}

\item{cndtn_series}{A column in df1 representing the conditioning time series to derive the multiple binary indicators}

\item{bin_method}{either, "level" - split time series into terciles only, or "both"  - split time series into terciles and a 6 month change indicator ("increase" or "decrease")}

\item{lb}{The look back period for draw-down assessment}

\item{pc}{The percent draw-down for binary market in/out indicator cutoff}

\item{fr}{The minimum forward return for binary market in/out indicator cutoff}
}
\value{
A ggplot object.
}
\description{
Make density plot of subsequent returns conditioned on multiple binary indicators derived from a reference time series
}
