% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_segmentation.R
\name{ts_segmentation}
\alias{ts_segmentation}
\title{Time series histogram and shading}
\usage{
ts_segmentation(df1, df2, date_idx, cndtn_series, bin_method)
}
\arguments{
\item{df1}{A dataframe containing the following columns:
\itemize{
\item date
\item market index values (labelled "close")
\item monthly forward market returns (labelled "fwd_rtn_m")
\item an indicator time series to be plotted and categorised into bins representing specific level and change values
}}

\item{df2}{A dataframe containing the time series to shade the S&P500 plot}

\item{date_idx}{The column of df1 which is the date index}

\item{cndtn_series}{A column name in df1 representing the conditioning time series to derive the multiple binary indicators}

\item{bin_method}{either, "level" - split time series into terciles only, or "both"  - split time series into terciles and a 6 month change indicator ("increase" or "decrease")}
}
\value{
A ggplot object.
}
\description{
Make density plot of subsequent returns conditioned on multiple binary indicators derived from reference time series
}
